relocate world model network to torch_utils/dreamer
