+++ ding/model/common/encoder.py
@@ -6,7 +6,8 @@ import torch
 import torch.nn as nn
 from torch.nn import functional as F
 
-from ding.torch_utils import ResFCBlock, ResBlock, Flatten, normed_linear, normed_conv2d, Conv2dSame, DreamerLayerNorm
+from ding.torch_utils import ResFCBlock, ResBlock, Flatten, normed_linear, normed_conv2d
+from ding.torch_utils.network.dreamer import Conv2dSame, DreamerLayerNorm
 from ding.utils import SequenceType
