+++ ding/world_model/model/networks.py
@@ -6,9 +6,7 @@ from torch import nn
 import torch.nn.functional as F
 from torch import distributions as torchd
 
-from ding.world_model.utils import weight_init, uniform_weight_init, OneHotDist, ContDist, SymlogDist, SampleDist, \
-    Bernoulli, TwoHotDistSymlog, UnnormalizedHuber, SafeTruncatedNormal, TanhBijector, static_scan
-from ding.torch_utils import MLP, DreamerLayerNorm
+from ding.torch_utils.network.dreamer import weight_init, uniform_weight_init, OneHotDist, ContDist, SymlogDist, static_scan, DreamerLayerNorm
 
 
 class RSSM(nn.Module):
@@ -366,167 +364,6 @@ class ConvDecoder(nn.Module):
         mean = x.reshape(features.shape[:-1] + self._shape)
         mean = mean.permute(0, 1, 3, 4, 2)
         return SymlogDist(mean)
-
-
-class DenseHead(nn.Module):
-
-    def __init__(
-        self,
-        inp_dim,  # config.dyn_stoch * config.dyn_discrete + config.dyn_deter
-        shape,  # (255,)
-        layer_num,
-        units,  # 512
-        act='SiLU',
-        norm='LN',
-        dist='normal',
-        std=1.0,
-        outscale=1.0,
-    ):
-        super(DenseHead, self).__init__()
-        self._shape = (shape, ) if isinstance(shape, int) else shape
-        if len(self._shape) == 0:
-            self._shape = (1, )
-        self._layer_num = layer_num
-        self._units = units
-        self._act = getattr(torch.nn, act)()
-        self._norm = norm
-        self._dist = dist
-        self._std = std
-
-        self.mlp = MLP(
-            inp_dim,
-            self._units,
-            self._units,
-            self._layer_num,
-            layer_fn=nn.Linear,
-            activation=self._act,
-            norm_type=self._norm
-        )
-        self.mlp.apply(weight_init)
-
-        self.mean_layer = nn.Linear(self._units, np.prod(self._shape))
-        self.mean_layer.apply(uniform_weight_init(outscale))
-
-        if self._std == "learned":
-            self.std_layer = nn.Linear(self._units, np.prod(self._shape))
-            self.std_layer.apply(uniform_weight_init(outscale))
-
-    def forward(self, features, dtype=None):
-        x = features
-        out = self.mlp(x)  # (batch, time, _units=512)
-        mean = self.mean_layer(out)  # (batch, time, 255)
-        if self._std == "learned":
-            std = self.std_layer(out)
-        else:
-            std = self._std
-        if self._dist == "normal":
-            return ContDist(torchd.independent.Independent(torchd.normal.Normal(mean, std), len(self._shape)))
-        if self._dist == "huber":
-            return ContDist(torchd.independent.Independent(UnnormalizedHuber(mean, std, 1.0), len(self._shape)))
-        if self._dist == "binary":
-            return Bernoulli(torchd.independent.Independent(torchd.bernoulli.Bernoulli(logits=mean), len(self._shape)))
-        if self._dist == "twohot_symlog":
-            return TwoHotDistSymlog(logits=mean)
-        raise NotImplementedError(self._dist)
-
-
-class ActionHead(nn.Module):
-
-    def __init__(
-        self,
-        inp_dim,
-        size,
-        layers,
-        units,
-        act=nn.ELU,
-        norm=nn.LayerNorm,
-        dist="trunc_normal",
-        init_std=0.0,
-        min_std=0.1,
-        max_std=1.0,
-        temp=0.1,
-        outscale=1.0,
-        unimix_ratio=0.01,
-    ):
-        super(ActionHead, self).__init__()
-        self._size = size
-        self._layers = layers
-        self._units = units
-        self._dist = dist
-        self._act = act
-        self._norm = norm
-        self._min_std = min_std
-        self._max_std = max_std
-        self._init_std = init_std
-        self._unimix_ratio = unimix_ratio
-        self._temp = temp() if callable(temp) else temp
-
-        pre_layers = []
-        for index in range(self._layers):
-            pre_layers.append(nn.Linear(inp_dim, self._units, bias=False))
-            pre_layers.append(norm(self._units, eps=1e-03))
-            pre_layers.append(act())
-            if index == 0:
-                inp_dim = self._units
-        self._pre_layers = nn.Sequential(*pre_layers)
-        self._pre_layers.apply(weight_init)
-
-        if self._dist in ["tanh_normal", "tanh_normal_5", "normal", "trunc_normal"]:
-            self._dist_layer = nn.Linear(self._units, 2 * self._size)
-            self._dist_layer.apply(uniform_weight_init(outscale))
-
-        elif self._dist in ["normal_1", "onehot", "onehot_gumbel"]:
-            self._dist_layer = nn.Linear(self._units, self._size)
-            self._dist_layer.apply(uniform_weight_init(outscale))
-
-    def __call__(self, features, dtype=None):
-        x = features
-        x = self._pre_layers(x)
-        if self._dist == "tanh_normal":
-            x = self._dist_layer(x)
-            mean, std = torch.split(x, 2, -1)
-            mean = torch.tanh(mean)
-            std = F.softplus(std + self._init_std) + self._min_std
-            dist = torchd.normal.Normal(mean, std)
-            dist = torchd.transformed_distribution.TransformedDistribution(dist, TanhBijector())
-            dist = torchd.independent.Independent(dist, 1)
-            dist = SampleDist(dist)
-        elif self._dist == "tanh_normal_5":
-            x = self._dist_layer(x)
-            mean, std = torch.split(x, 2, -1)
-            mean = 5 * torch.tanh(mean / 5)
-            std = F.softplus(std + 5) + 5
-            dist = torchd.normal.Normal(mean, std)
-            dist = torchd.transformed_distribution.TransformedDistribution(dist, TanhBijector())
-            dist = torchd.independent.Independent(dist, 1)
-            dist = SampleDist(dist)
-        elif self._dist == "normal":
-            x = self._dist_layer(x)
-            mean, std = torch.split(x, [self._size] * 2, -1)
-            std = (self._max_std - self._min_std) * torch.sigmoid(std + 2.0) + self._min_std
-            dist = torchd.normal.Normal(torch.tanh(mean), std)
-            dist = ContDist(torchd.independent.Independent(dist, 1))
-        elif self._dist == "normal_1":
-            x = self._dist_layer(x)
-            dist = torchd.normal.Normal(mean, 1)
-            dist = ContDist(torchd.independent.Independent(dist, 1))
-        elif self._dist == "trunc_normal":
-            x = self._dist_layer(x)
-            mean, std = torch.split(x, [self._size] * 2, -1)
-            mean = torch.tanh(mean)
-            std = 2 * torch.sigmoid(std / 2) + self._min_std
-            dist = SafeTruncatedNormal(mean, std, -1, 1)
-            dist = ContDist(torchd.independent.Independent(dist, 1))
-        elif self._dist == "onehot":
-            x = self._dist_layer(x)
-            dist = OneHotDist(x, unimix_ratio=self._unimix_ratio)
-        elif self._dist == "onehot_gumble":
-            x = self._dist_layer(x)
-            temp = self._temp
-            dist = ContDist(torchd.gumbel.Gumbel(x, 1 / temp))
-        else:
-            raise NotImplementedError(self._dist)
-        return dist
 
 
 class GRUCell(nn.Module):
