+++ ding/world_model/utils.py
@@ -1,10 +1,5 @@
 from easydict import EasyDict
 from typing import Callable
-import numpy as np
-import torch
-from torch import nn
-from torch.nn import functional as F
-from torch import distributions as torchd
 
 
 def get_rollout_length_scheduler(cfg: EasyDict) -> Callable[[int], int]:
@@ -28,326 +23,3 @@ def get_rollout_length_scheduler(cfg: EasyDict) -> Callable[[int], int]:
         return lambda x: cfg.rollout_length
     else:
         raise KeyError("not implemented key: {}".format(cfg.type))
-
-
-def symlog(x):
-    return torch.sign(x) * torch.log(torch.abs(x) + 1.0)
-
-
-def symexp(x):
-    return torch.sign(x) * (torch.exp(torch.abs(x)) - 1.0)
-
-
-class SampleDist:
-
-    def __init__(self, dist, samples=100):
-        self._dist = dist
-        self._samples = samples
-
-    @property
-    def name(self):
-        return 'SampleDist'
-
-    def __getattr__(self, name):
-        return getattr(self._dist, name)
-
-    def mean(self):
-        samples = self._dist.sample(self._samples)
-        return torch.mean(samples, 0)
-
-    def mode(self):
-        sample = self._dist.sample(self._samples)
-        logprob = self._dist.log_prob(sample)
-        return sample[torch.argmax(logprob)][0]
-
-    def entropy(self):
-        sample = self._dist.sample(self._samples)
-        logprob = self.log_prob(sample)
-        return -torch.mean(logprob, 0)
-
-
-class OneHotDist(torchd.one_hot_categorical.OneHotCategorical):
-
-    def __init__(self, logits=None, probs=None, unimix_ratio=0.0):
-        if logits is not None and unimix_ratio > 0.0:
-            probs = F.softmax(logits, dim=-1)
-            probs = probs * (1.0 - unimix_ratio) + unimix_ratio / probs.shape[-1]
-            logits = torch.log(probs)
-            super().__init__(logits=logits, probs=None)
-        else:
-            super().__init__(logits=logits, probs=probs)
-
-    def mode(self):
-        _mode = F.one_hot(torch.argmax(super().logits, axis=-1), super().logits.shape[-1])
-        return _mode.detach() + super().logits - super().logits.detach()
-
-    def sample(self, sample_shape=(), seed=None):
-        if seed is not None:
-            raise ValueError('need to check')
-        sample = super().sample(sample_shape)
-        probs = super().probs
-        while len(probs.shape) < len(sample.shape):
-            probs = probs[None]
-        sample += probs - probs.detach()
-        return sample
-
-
-class TwoHotDistSymlog():
-
-    def __init__(self, logits=None, low=-20.0, high=20.0, device='cpu'):
-        self.logits = logits
-        self.probs = torch.softmax(logits, -1)
-        self.buckets = torch.linspace(low, high, steps=255).to(device)
-        self.width = (self.buckets[-1] - self.buckets[0]) / 255
-
-    def mean(self):
-        print("mean called")
-        _mean = self.probs * self.buckets
-        return symexp(torch.sum(_mean, dim=-1, keepdim=True))
-
-    def mode(self):
-        _mode = self.probs * self.buckets
-        return symexp(torch.sum(_mode, dim=-1, keepdim=True))
-
-    # Inside OneHotCategorical, log_prob is calculated using only max element in targets
-    def log_prob(self, x):
-        x = symlog(x)
-        # x(time, batch, 1)
-        below = torch.sum((self.buckets <= x[..., None]).to(torch.int32), dim=-1) - 1
-        above = len(self.buckets) - torch.sum((self.buckets > x[..., None]).to(torch.int32), dim=-1)
-        below = torch.clip(below, 0, len(self.buckets) - 1)
-        above = torch.clip(above, 0, len(self.buckets) - 1)
-        equal = (below == above)
-
-        dist_to_below = torch.where(equal, 1, torch.abs(self.buckets[below] - x))
-        dist_to_above = torch.where(equal, 1, torch.abs(self.buckets[above] - x))
-        total = dist_to_below + dist_to_above
-        weight_below = dist_to_above / total
-        weight_above = dist_to_below / total
-        target = (
-            F.one_hot(below, num_classes=len(self.buckets)) * weight_below[..., None] +
-            F.one_hot(above, num_classes=len(self.buckets)) * weight_above[..., None]
-        )
-        log_pred = self.logits - torch.logsumexp(self.logits, -1, keepdim=True)
-        target = target.squeeze(-2)
-
-        return (target * log_pred).sum(-1)
-
-    def log_prob_target(self, target):
-        log_pred = super().logits - torch.logsumexp(super().logits, -1, keepdim=True)
-        return (target * log_pred).sum(-1)
-
-
-class SymlogDist():
-
-    def __init__(self, mode, dist='mse', agg='sum', tol=1e-8, dim_to_reduce=[-1, -2, -3]):
-        self._mode = mode
-        self._dist = dist
-        self._agg = agg
-        self._tol = tol
-        self._dim_to_reduce = dim_to_reduce
-
-    def mode(self):
-        return symexp(self._mode)
-
-    def mean(self):
-        return symexp(self._mode)
-
-    def log_prob(self, value):
-        assert self._mode.shape == value.shape
-        if self._dist == 'mse':
-            distance = (self._mode - symlog(value)) ** 2.0
-            distance = torch.where(distance < self._tol, 0, distance)
-        elif self._dist == 'abs':
-            distance = torch.abs(self._mode - symlog(value))
-            distance = torch.where(distance < self._tol, 0, distance)
-        else:
-            raise NotImplementedError(self._dist)
-        if self._agg == 'mean':
-            loss = distance.mean(self._dim_to_reduce)
-        elif self._agg == 'sum':
-            loss = distance.sum(self._dim_to_reduce)
-        else:
-            raise NotImplementedError(self._agg)
-        return -loss
-
-
-class ContDist:
-
-    def __init__(self, dist=None):
-        super().__init__()
-        self._dist = dist
-        self.mean = dist.mean
-
-    def __getattr__(self, name):
-        return getattr(self._dist, name)
-
-    def entropy(self):
-        return self._dist.entropy()
-
-    def mode(self):
-        return self._dist.mean
-
-    def sample(self, sample_shape=()):
-        return self._dist.rsample(sample_shape)
-
-    def log_prob(self, x):
-        return self._dist.log_prob(x)
-
-
-class Bernoulli:
-
-    def __init__(self, dist=None):
-        super().__init__()
-        self._dist = dist
-        self.mean = dist.mean
-
-    def __getattr__(self, name):
-        return getattr(self._dist, name)
-
-    def entropy(self):
-        return self._dist.entropy()
-
-    def mode(self):
-        _mode = torch.round(self._dist.mean)
-        return _mode.detach() + self._dist.mean - self._dist.mean.detach()
-
-    def sample(self, sample_shape=()):
-        return self._dist.rsample(sample_shape)
-
-    def log_prob(self, x):
-        _logits = self._dist.base_dist.logits
-        log_probs0 = -F.softplus(_logits)
-        log_probs1 = -F.softplus(-_logits)
-
-        return log_probs0 * (1 - x) + log_probs1 * x
-
-
-class UnnormalizedHuber(torchd.normal.Normal):
-
-    def __init__(self, loc, scale, threshold=1, **kwargs):
-        super().__init__(loc, scale, **kwargs)
-        self._threshold = threshold
-
-    def log_prob(self, event):
-        return -(torch.sqrt((event - self.mean) ** 2 + self._threshold ** 2) - self._threshold)
-
-    def mode(self):
-        return self.mean
-
-
-class SafeTruncatedNormal(torchd.normal.Normal):
-
-    def __init__(self, loc, scale, low, high, clip=1e-6, mult=1):
-        super().__init__(loc, scale)
-        self._low = low
-        self._high = high
-        self._clip = clip
-        self._mult = mult
-
-    def sample(self, sample_shape):
-        event = super().sample(sample_shape)
-        if self._clip:
-            clipped = torch.clip(event, self._low + self._clip, self._high - self._clip)
-            event = event - event.detach() + clipped.detach()
-        if self._mult:
-            event *= self._mult
-        return event
-
-
-class TanhBijector(torchd.Transform):
-
-    def __init__(self, validate_args=False, name='tanh'):
-        super().__init__()
-
-    def _forward(self, x):
-        return torch.tanh(x)
-
-    def _inverse(self, y):
-        y = torch.where((torch.abs(y) <= 1.), torch.clamp(y, -0.99999997, 0.99999997), y)
-        y = torch.atanh(y)
-        return y
-
-    def _forward_log_det_jacobian(self, x):
-        log2 = torch.math.log(2.0)
-        return 2.0 * (log2 - x - torch.softplus(-2.0 * x))
-
-
-def static_scan(fn, inputs, start):
-    last = start  # {logit:[batch_size, self._stoch, self._discrete], stoch:[batch_size, self._stoch, self._discrete], deter:[batch_size, self._deter]}
-    indices = range(inputs[0].shape[0])
-    flag = True
-    for index in indices:
-        inp = lambda x: (_input[x] for _input in inputs)  # inputs:(action:(time, batch, 6), embed:(time, batch, 4096))
-        last = fn(last, *inp(index))  # post, prior
-        if flag:
-            if type(last) == type({}):
-                outputs = {key: value.clone().unsqueeze(0) for key, value in last.items()}
-            else:
-                outputs = []
-                for _last in last:
-                    if type(_last) == type({}):
-                        outputs.append({key: value.clone().unsqueeze(0) for key, value in _last.items()})
-                    else:
-                        outputs.append(_last.clone().unsqueeze(0))
-            flag = False
-        else:
-            if type(last) == type({}):
-                for key in last.keys():
-                    outputs[key] = torch.cat([outputs[key], last[key].unsqueeze(0)], dim=0)
-            else:
-                for j in range(len(outputs)):
-                    if type(last[j]) == type({}):
-                        for key in last[j].keys():
-                            outputs[j][key] = torch.cat([outputs[j][key], last[j][key].unsqueeze(0)], dim=0)
-                    else:
-                        outputs[j] = torch.cat([outputs[j], last[j].unsqueeze(0)], dim=0)
-    if type(last) == type({}):
-        outputs = [outputs]
-    return outputs
-
-
-def weight_init(m):
-    if isinstance(m, nn.Linear):
-        in_num = m.in_features
-        out_num = m.out_features
-        denoms = (in_num + out_num) / 2.0
-        scale = 1.0 / denoms
-        std = np.sqrt(scale) / 0.87962566103423978
-        nn.init.trunc_normal_(m.weight.data, mean=0.0, std=std, a=-2.0, b=2.0)
-        if hasattr(m.bias, 'data'):
-            m.bias.data.fill_(0.0)
-    elif isinstance(m, nn.Conv2d) or isinstance(m, nn.ConvTranspose2d):
-        space = m.kernel_size[0] * m.kernel_size[1]
-        in_num = space * m.in_channels
-        out_num = space * m.out_channels
-        denoms = (in_num + out_num) / 2.0
-        scale = 1.0 / denoms
-        std = np.sqrt(scale) / 0.87962566103423978
-        nn.init.trunc_normal_(m.weight.data, mean=0.0, std=std, a=-2.0, b=2.0)
-        if hasattr(m.bias, 'data'):
-            m.bias.data.fill_(0.0)
-    elif isinstance(m, nn.LayerNorm):
-        m.weight.data.fill_(1.0)
-        if hasattr(m.bias, 'data'):
-            m.bias.data.fill_(0.0)
-
-
-def uniform_weight_init(given_scale):
-
-    def f(m):
-        if isinstance(m, nn.Linear):
-            in_num = m.in_features
-            out_num = m.out_features
-            denoms = (in_num + out_num) / 2.0
-            scale = given_scale / denoms
-            limit = np.sqrt(3 * scale)
-            nn.init.uniform_(m.weight.data, a=-limit, b=limit)
-            if hasattr(m.bias, 'data'):
-                m.bias.data.fill_(0.0)
-        elif isinstance(m, nn.LayerNorm):
-            m.weight.data.fill_(1.0)
-            if hasattr(m.bias, 'data'):
-                m.bias.data.fill_(0.0)
-
-    return f
